Fix double-sanitization issue
When a user logs in with a special character in his username, he
recieves a sanitized username. However, when the user tries to for
example create a room, his sanitized username is sanitized again,
causing a authentication error. This commit fixes this issue, by
first desanitizing the payload before sanitizing it, preventing a
re-sanitization