Fix redis HSET error
When inspecting the logs, I found this error:

node_redis: Deprecated: The HSET command contains a "undefined" argument.
This is converted to a "undefined" string now and will return an error from v.3.0 on.
Please handle this in your code to make sure everything works as you intended it to.

The node app was trying to save the room's turn as a backup if
players should rejoin. However room.turn might be "undefined",
causing the error above. This commit adds a fix, by setting an
empty string if the room turn is falsey.