Ignore "room not found" error in onclose.js
When closing a room and kicking the players an error occurs that
the room doesn't exist. This is because the room is deleted
through the API before the game-server is notified through redis
pub/sub to kick the players. The error should be ignored as it is
unavoidable with the current architecture.