Remove room from openrooms when not found
A room might have been removed/expired but is still available
through the openrooms set. If this is the case and user tries to
connect to this room, the connection fails as the room is not
found. In this case the room should alse be removed from the
openrooms set.