Log user in when joining a private room
When a user receives an invitation link to a private room he
currently has to visit the game website first, log in and then
paste the invitation link.
A user is now redirected to the login/home page when joining a room
without being logged in. After login the user is then redirected
back to the room.