Rewrite db connection setup
Use the new node-redis package, including the new es6 import
syntax. The new redis methods are async by default, so bluebird is
no longer needed.
node-redis uses a connection pool which conflicts with redis monitor
mode in development, because monitor mode takes over the connection.
As a workaround, a redis-cli instance in monitor mode is launched in
the background and streams back stdin into the node process.